Include fcppt/config.hpp.

+++ src/plugins/opengl/xf86vmode/src/calc_refresh_rate.cpp
@@ -21,10 +21,22 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include "../calc_refresh_rate.hpp"
 #include <fcppt/math/round_div_int.hpp>
 
-unsigned sge::opengl::xf86vmode::calc_refresh_rate(
-	XF86VidModeModeInfo const &mode)
+unsigned
+sge::opengl::xf86vmode::calc_refresh_rate(
+	XF86VidModeModeInfo const &mode
+)
 {
-	return fcppt::math::round_div_int(
-		1000 * mode.dotclock,
-		static_cast<unsigned>(mode.htotal * mode.vtotal));
+	return
+		fcppt::math::round_div_int(
+			static_cast<
+				unsigned
+			>(
+				1000 * mode.dotclock
+			),
+			static_cast<
+				unsigned
+			>(
+				mode.htotal * mode.vtotal
+			)
+		);
 }
